renamed zendesk folder to graphql for clarity
