+++ .circleci/config.yml
@@ -26,7 +26,18 @@ jobs:
     steps:
       - checkout:
           path: ~/Deskmate
-      - run: cd server
+      - go/load-cache
+      - go/mod-download-cached
+      - go/save-cache
+      - run: go vet
+      - run: go test -v ./...
+  test-graphql:
+    working_directory: ~/Deskmate/zendesk
+    docker:
+      - image: circleci/golang
+    steps:
+      - checkout:
+          path: ~/Deskmate
       - go/load-cache
       - go/mod-download-cached
       - go/save-cache
@@ -39,3 +50,4 @@ workflows:
     jobs:
       - test-client
       - test-server
+      - test-graphql
