+++ graphql/README.md
@@ -1,0 +1,2 @@
+# Deskmate/graphql
+The API backbone for the Deskmate project
