+++ graphql/gqlgen.yml
