+++ graphql/resolver/resolver.go
@@ -3,7 +3,7 @@ package resolver
 // THIS CODE IS A STARTING POINT ONLY. IT WILL NOT BE UPDATED WITH SCHEMA CHANGES.
 
 import (
-	"github.com/tylerconlee/Deskmate/zendesk/graph"
+	"github.com/tylerconlee/Deskmate/graphql/graph"
 )
 
 type Resolver struct{}
