--- zendesk/.circleci/config.yml
@@ -1,34 +1,0 @@
-version: 2.1
-orbs:
-  go: circleci/go@0.2.0
-executors:
-  golang: go/default
-jobs:
-  build:
-    working_directory: /go/src/github.com/tylerconlee/Deskmate/zendesk
-    executor: golang   
-    steps:
-      - checkout
-      - go/load-cache
-      - go/mod-download
-      - go/save-cache
-  test:
-    working_directory: /go/src/github.com/tylerconlee/Deskmate/zendesk
-    executor: golang
-    parallelism: 2
-    steps:
-      - checkout
-      - go/load-cache
-      - run: go get -u github.com/jstemmer/go-junit-report 
-      - run: mkdir go-test 
-      - run: go test -v $(go list ./... | grep -v /vendor/) | go-junit-report > go-test/results.xml
-      - store_test_results:
-           path: go-test/
-workflows:
-  version: 2
-  build_and_test:
-    jobs:
-      - build
-      - test:
-          requires:
-            - build
