--- zendesk/README.md
@@ -1,2 +1,0 @@
-# Deskmate/zendesk
-The API backbone for the Slab project
